backend: move service to utils
