+++ portal-backend/src/utils/util.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { MailerService } from './services/mailer.services';
+import { JwtSigningService } from './services/jwt-signing.services';
+import { CertificateModule } from '../certificates/certificate.module';
 
 @Module({
-  imports: [],
-  providers: [MailerService],
-  exports: [MailerService],
+  imports: [CertificateModule],
+  providers: [MailerService, JwtSigningService],
+  exports: [MailerService, JwtSigningService],
 })
 export class UtilsModule {}
